# ZerochSharp/ReactClient/src/models/thread.ts
import { Response } from './response';

export interface Thread {
  title: string;
  threadId: number;
  key?: number; 
  created: string;
  boardKey?: string; 
  author: string;
  modified: string;
  responseCount: number;
  datKey?: number;
  influence: number;
  responses?: Response[];
  archived: boolean;
  stopped: boolean;
}