# docs/api_documentation.yml
openapi: 3.0.0
info:
  title: Zeroch Sharp API
  description: API definitions of Zeroch Sharp RESTful API
  license:
    name: MIT
  version: "0.1"
servers:
  - url: https://bbs2.nreopigs.work/api
    description: Development Demo Server (nightly)
...
paths:
  /boards:
    get:
      summary: Get All boards.
      description: Get All boards in server.
      parameters: []
      tags: 
        - Boards
      responses:
        "200":
          description: A JSON array of Board model
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Board'
                example: 
                  - id: 1
                    boardName: News7VIP
                    boardKey: news7vip
                    boardSumbaTime: 30
                    boardSubTitle: This is a sample
                    boardDefaultName: Noname of news7vip
                    boardDeleteName: aboned
                    autoArchivingPredicates: [Count < 3 & Influence < 10]
                  - id: 2
                    boardName: Tea for two
                    boardKey: coffeehouse
                    boardSubTitle: null
                    boardSumbaTime: 30
                    boardDefaultName: Noname of coffeehouse
                    autoArchivingPredicates: []
        "404":
          description: Not found such a board
  /boards/{boardKey}:
    parameters: 
      - name: boardKey
        in: path
        description: board key
        required: true
        schema:
          type: string
    
    get:
      tags: 
      - Boards
      summary: Get board with thread list.
      description: Get thread list.
      responses:
        '200':
          description: A Json object of Board model
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Board'
              example: 
                  id: 1
                  boardKey: news7vip
                  boardName: UraVIP
                  boardDefaultName: Nameless people
                  boardSambaTime: 30
                  boardSubTitle: null
                  boardDeleteName: "Abone"
                  autoArchivingPredicates: [Count < 3 & Influence < 10]
                  children: [
                            {
                              id: 1,
                              author: gSr1De2,
                              created: "2020-01-23T15:11:39.257022",
                              boardKey: news7vip,
                              title: "AAAhhhh...!!!",
                              modified: "2020-01-26T17:36:23.596691",
                              responseCount: 11,
                              influence: 2.4114,
                              datKey: 1579759899
                            }
                         ]
        '404':
          description: Not found such a thread
    post:
      tags: 
        - Boards
      summary: Post thread to same name board
      description: Post thread to Board.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                title:
                  type: string
                response: 
                  type: object
                  properties:
                    name: 
                      type: string
                    mail:
                      type: string
                    body:
                      type: string
            example:
              title: "Hello, Zeroch Sharp!"
              response:
                name: ""
                mail: "sage"
                body: "Hello World!"
      responses:
        '200':
          description: Post thread correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
              example:
                  id: 1
                  boardKey: news7vip
                  boardName: UraVIP
                  boardDefaultName: Nameless people
                  boardSambaTime: 30
                  boardSubTitle: null
                  boardDeleteName: "Abone"
                  autoArchivingPredicates: [Count < 3 & Influence < 10]
                  children: [
                            {
                              id: 1,
                              author: gSr1De2,
                              created: "2020-01-23T15:11:39.257022",
                              boardKey: news7vip,
                              title: "Hello, Zeroch Sharp!",
                              modified: "2020-01-26T17:36:23.596691",
                              responseCount: 11,
                              influence: 2.4114,
                              datKey: 1579759899
                            }
                         ]
  /boards/{boardKey}/billboard:
    parameters: 
      - name: boardKey
        schema:
          type: string
        in: path
        required: true
        description: board key
    get:
      tags:
        - Boards
      summary: Get billboard image-path of specific board
      description: Get billboard image-path
      responses:
        '200':
          description: a path of image
          content:
            application/json:
              schema:
                type: object
                properties:
                  path:
                    type: string
        '404':
          description: not found board or billboard image
  /boards/{boardKey}/localrule:
    parameters:
      - name: boardKey
        schema:
          type: string
        required: true
        in: path
        description: board key
    get:
      tags:
        - Boards
      summary: Get localrule of specific board
      description: Get localrule
      responses:
        '200':
          description: full text of local rule
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: string
        '404':
          description: not found such a board
  /boards/{boardKey}/{threadId}:
    parameters: 
    - name: boardKey
      in: path
      description: board key
      required: true
      schema:
        type: string
    - name: threadId
      in: path
      description: thread Id
      required: true
      schema:
        type: string
    get:
      summary: Get Thread with responses.
      description: Get thread.
      tags:
        - Boards
      responses:
        '200':
          description: A json object of Thread model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Thread'
              example: 
                - id: 1
                  title: "AAAhhhh...!!!"
                  author: gEfec3fFI
                  created: "2020-01-23T15:11:39.257022"
                  modified: "2020-01-26T17:36:23.596691"
                  responseCount: 2
                  influence: 234.23
                  boardKey: news7vip
                  datKey: 1579759899
                  responses:
                    - id: 1
                      author: gEfec3fFI
                      created: "2020-01-23T15:11:39.257022"
                      threadId: 1
                      hostAddress: 127.0.0.1
                      isAboned: false
                      body: "Ohhhhhh...!!!!!"
                      mail: ""
                      name: ""
                    - id: 2
                      author: gEfec3fFI
                      created: "2020-01-23T15:11:39.257022"
                      threadId: 1
                      hostAddress: 127.0.0.1
                      isAboned: true
                      body: "aboned"
                      mail: "aboned"
                      name: "aboned"
    post:
      tags:
        - Boards
      summary: Post response
      description: Post Response to same thread
      requestBody:
        content: 
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name: 
                  type: string
                mail:
                  type: string
                body:
                  type: string
            example:
              name: "Hello"
              mail: ""
              body: "Hello, world!"
      responses:
        '200':
          description: Post response correctly
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Response'
              example:
                id: 4
                author: "fgEtrYj3rd"
                created: "2020-02-04T17:00:24.3568815+09:00"
                threadId: 3
                isAboned: false
                body: "Hello, world!"
                mail: ""
                name: "Hello"
    delete:
      tags:
        - Boards
      summary: Delete Thread
      description: Delete thread from board
      parameters:
        - name: isRemove
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: successful of remove thread
        '404':
          description: not found such a thread
        '401':
          description: no authority to delete thread
    patch:
      summary: Patch any attribute to thread
      description: Patch attributes.
      tags:
        - Boards
      responses:
        '200':
          description: Successful to patch attributes.
        '401':
          description: No authorization to patch attribute to thread
        '404':
          description: cannot find such a thread
  /boards/{boardKey}/{threadId}/{responseId}:
    parameters:
      - name: boardKey
        in: path
        required: true
        schema:
          type: string
        description: board key
      - name: threadId
        in: path
        required: true
        schema:
          type: string
        description: thread id
      - name: responseId
        in: path
        required: true
        description: response id
        schema:
          type: string
    delete:
      tags:
        - Boards
      description: Delete response from thread
      parameters:
      - name: isRemove
        in: query
        schema:
          type: boolean
          default: false
      responses:
        '200': 
          description: success deleting response
        '400':
          description: not found such a response
        '401':
          description: no authority to delete response
    put:
      tags:
        - Boards
      description: Edit response
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Response'
      responses:
        '200':
          description: success editing response
        '401':
          description: no authority to edit response
...
  /boards/{boardKey}/setting:
    parameters:
      - name: boardKey
        description: board key
        in: path
        schema:
          type: string
        required: true
    post:
      tags:
        - Boards setting
      description: setting of specific board
      summary: Post board setting
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                boardName:
                  type: string
                  nullable: true
                boardDefaultName:
                  type: string
                  nullable: true
                boardSubTitle:
                  type: string
                  nullable: true
                autoArchivingPredicates:
                  type: array
                  items:
                    type: string
                  nullable: true
      responses:
        '401':
          description: no authority to change board setting
        '200':
          description: successful to change board setting
        '404':
          description: cannot find specific board
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
  schemas:
    Board:
      type: object
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        boardName:
          type: string
        boardKey:
          type: string
        boardSumbaTime:
          type: integer
          format: int64
        boardSubTitle:
          type: string
          nullable: true
        boardDefaultName:
          type: string
        boardDeleteName:
          type: string
        children:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Thread'
        autoArchivingPredicates:
          type: array
          items:
           type: string
          nullable: true
      additionalProperties: false
    Thread:
      type: object
      properties:
        threadId:
          type: integer
          format: int64
          minimum: 1
        title:
          type: string
        created:
          type: string
        modified:
          type: string
        responseCount:
          type: integer
          format: int64
        influence:
          type: number
          format: double
        boardKey:
          type: string
        author:
          type: string
        datKey:
          type: integer
          format: int64
        responses:
          type: array
          items:
            $ref: '#/components/schemas/Response'
          nullable: true
      additionalProperties: false
    Response:
      type: object
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        author:
          type: string
        created:
          type: string
        threadId:
          type: integer
          format: int64
        hostAddress:
          type: string
          nullable: true
        isAboned:
          type: boolean
        body:
          type: string
        mail:
          type: string
        name: 
          type: string
      additionalProperties: false
...